updated more css
