+++ frontend/src/app/globals.css
@@ -2,7 +2,7 @@
   --bg-primary: #eaf2fc;
   --bg-light: #ffffff;
   --bg-dark: #c8d9f0;
-  --bg-content: #3a4f6e;
+  --bg-content: #455e85;
   --text-primary: #1e2837; 
   --text-secondary: #4b5c77;
   --bg-accent: #2d88f0;
