+++ frontend/src/components/GameDisplays/DashGameGallery/DashGameGallery.module.css
@@ -3,7 +3,7 @@
     justify-content: center;
     align-items: center;
     padding: 80px;
-    color: var(--text-primary);
+    color: white;
     font-family: Arial, sans-serif;
 }
     
@@ -20,19 +20,19 @@
     transition: transform 0.3s ease-in-out, filter 0.3s ease-in-out, box-shadow 0.3s ease-in-out;
     filter: brightness(0.9);
     position: relative;
-    }
+}
     
-    .gameContainer:nth-child(even) {
-    background-color: rgba(72, 72, 72, 0.6);
-    transform: scale(1.05);
-    }
-    
-    .gameContainer:hover {
-    transform: scale(1.075) translateY(-5px);
-    filter: brightness(1.03);
-    box-shadow: 0 15px 30px rgba(0, 0, 0, 0.7), 0 0 15px rgba(98, 153, 240, 0.3);
-    animation: subtleRotate 0.5s ease-in-out;
-    }
+.gameContainer:nth-child(even) {
+  background-color: rgba(72, 72, 72, 0.6);
+  transform: scale(1.05);
+}
+
+.gameContainer:hover {
+  transform: scale(1.075) translateY(-5px);
+  filter: brightness(1.03);
+  box-shadow: 0 15px 30px rgba(0, 0, 0, 0.7), 0 0 15px rgba(98, 153, 240, 0.3);
+  animation: subtleRotate 0.5s ease-in-out;
+}
     
 .gallery img {
     border: 1px solid white;
@@ -56,6 +56,5 @@
 .gameCategory {
     margin-top: 5px;
     font-size: 1.4rem;
-    color: var(--text-primary);
     font-weight: bold;
 }
