+++ frontend/src/components/GameDisplays/GameTable/GameTable.jsx
@@ -30,11 +30,18 @@ const GameTable = ({ games }) => {
                   ? `${game.hltb_score} Hrs`
                   : 'Unknown'}
               </td>
-              <td>{game.boil_score !== null && game.boil_score !== undefined
-                  ? game.boil_score
-                  : 'N/A'}</td>
               <td>
-                <a href={'https://store.steampowered.com/app/' + game.game_id}><img src="https://upload.wikimedia.org/wikipedia/commons/thumb/8/83/Steam_icon_logo.svg/512px-Steam_icon_logo.svg.png" className={styles.steamImg}/></a>
+                {game.boil_score !== null && game.boil_score !== undefined
+                  ? game.boil_score
+                  : 'N/A'}
+              </td>
+              <td>
+                <a href={'https://store.steampowered.com/app/' + game.game_id}>
+                  <img
+                    src='https://upload.wikimedia.org/wikipedia/commons/thumb/8/83/Steam_icon_logo.svg/512px-Steam_icon_logo.svg.png'
+                    className={styles.steamImg}
+                  />
+                </a>
               </td>
               <td>
                 <input type='checkbox' />
