+++ frontend/src/components/GameDisplays/TopRatedGames/TopRatedGames.module.css
@@ -33,7 +33,7 @@
   top: 5%;
   right: 65%;
   background-color: rgba(0, 0, 0, 0.5);
-  color: var(--text-primary);
+  color: white;
   font-family: Arial, sans-serif;
   font-size: 18px;
   font-weight: bold;
@@ -94,7 +94,7 @@
   top: 50%; /* Center vertically */
   left: 50%; /* Center horizontally */
   transform: translate(-50%, -50%); /* Adjust for the element's own size to truly center */
-  color: var(--text-primary);
+  color: white;
   font-family: Arial, sans-serif;
   font-size: 18px;
   font-weight: bold;
