+++ frontend/src/components/SteamComponents/SteamIdDisplay.module.css
@@ -87,7 +87,7 @@
 
 .steamName {
     font-size: 16px;
-    color: var(--text-primary);
+    color: white;
     margin: 4px 0;
 }
