--- frontend/src/providers/ThemeInitializer 2.jsx
@@ -1,29 +1,0 @@
-'use client'
-
-import { useEffect } from 'react'
-import { useTheme } from 'next-themes'
-import axios from 'axios'
-
-export function ThemeInitializer() {
-  const { theme, setTheme } = useTheme()
-
-  useEffect(() => {
-    async function fetchUserPreference() {
-      try {
-        const res = await axios.get(
-          process.env.NEXT_PUBLIC_BACKEND + '/themepreference',
-          { withCredentials: true }
-        )
-
-        console.log(res.data)
-
-        setTheme(res.data.preference == 0 ? 'dark' : 'light')
-      } catch (err) {
-        console.error(err)
-      }
-    }
-    fetchUserPreference()
-  }, [])
-
-  return null
-}
